refactor(tools): split read.rs into read/ module with pdf.rs and image.rs
read.rs was 943 LOC with PDF extraction, image processing, and the core
ReadTool all in one file. Future formats (.ipynb, .docx) would keep
growing it.

Refactored to a module directory:
 - read/mod.rs (510 LOC) — ReadTool, render_bytes_to_payload, tests
 - read/pdf.rs (153 LOC) — parse_page_range, read_pdf
 - read/image.rs (75 LOC) — is_image_extension, read_image

Also extracted the 990-line contract_tests block from fs/mod.rs (1221
LOC → 234 LOC) into a dedicated fs/contract_tests.rs file.

Skipped splitting shared.rs and grep.rs — their internals are tightly
coupled (FsState + FsPathPolicy + error helpers + binary detection all
cross-reference each other; grep's Sink impls are only used within
grep_impl). Splitting would add import boilerplate without readability
gain.

131 fs tests pass, clippy clean, gateway builds.

Entire-Checkpoint: 1c9addc03229
