+++ crates/tools/src/fs/read/mod.rs
@@ -4,6 +4,12 @@
 //! and `limit`. Returns a structured payload with `content` (cat -n style),
 //! `total_lines`, and `truncated` flags so the LLM can tell whether it has
 //! the full file.
+//!
+//! Format-specific dispatchers live in submodules so new formats (e.g.
+//! `.ipynb`, `.docx`) can be added without growing this file.
+
+pub(crate) mod image;
+pub(crate) mod pdf;
 
 use {
     async_trait::async_trait,
@@ -480,7 +486,7 @@ impl AgentTool for ReadTool {
 
         // Image dispatch.
         if is_image_extension(&lower) {
-            return match read_image(file_path).await {
+            return match image::read_image(file_path).await {
                 Ok(value) => Ok(value),
                 Err(e) => {
                     #[cfg(feature = "metrics")]
@@ -509,205 +515,7 @@ impl AgentTool for ReadTool {
     }
 }
 
-/// Image extensions that get special handling (resize + base64) instead
-/// of binary rejection. Matches Claude Code's `UYq` set.
-const IMAGE_EXTENSIONS: &[&str] = &["png", "jpg", "jpeg", "gif", "webp"];
-
-fn is_image_extension(lower_path: &str) -> bool {
-    IMAGE_EXTENSIONS.iter().any(|ext| lower_path.ends_with(ext))
-}
-
-/// Read an image file, optimize it for LLM consumption (resize to fit
-/// within dimension limits, compress), and return a structured payload
-/// with base64 data + original/final dimensions.
-///
-/// Uses `moltis_media::image_ops::optimize_for_llm` which already
-/// handles Lanczos3 resize, transparency detection (PNG for RGBA,
-/// JPEG otherwise), and progressive quality reduction.
-async fn read_image(file_path: &str) -> Result<Value> {
-    require_absolute(file_path, "file_path")?;
-
-    let path = std::path::PathBuf::from(file_path);
-    if !fs::try_exists(&path).await.unwrap_or(false) {
-        return Ok(json!({
-            "kind": "not_found",
-            "file_path": file_path,
-            "error": "file does not exist",
-            "detail": "",
-        }));
-    }
-
-    let bytes = fs::read(&path)
-        .await
-        .map_err(|e| Error::message(format!("failed to read image '{file_path}': {e}")))?;
-
-    let path_owned = file_path.to_string();
-    tokio::task::spawn_blocking(move || -> Result<Value> {
-        match moltis_media::image_ops::optimize_for_llm(&bytes, None) {
-            Ok(optimized) => {
-                use base64::{Engine as _, engine::general_purpose::STANDARD as BASE64};
-                Ok(json!({
-                    "kind": "image",
-                    "file_path": path_owned,
-                    "media_type": optimized.media_type,
-                    "original_width": optimized.original_width,
-                    "original_height": optimized.original_height,
-                    "final_width": optimized.final_width,
-                    "final_height": optimized.final_height,
-                    "was_resized": optimized.was_resized,
-                    "bytes": optimized.data.len(),
-                    "base64": BASE64.encode(&optimized.data),
-                }))
-            },
-            Err(e) => Ok(json!({
-                "kind": "image_error",
-                "file_path": path_owned,
-                "error": format!("failed to process image: {e}"),
-                "detail": "The file may be corrupted or use an unsupported image format.",
-            })),
-        }
-    })
-    .await
-    .map_err(|e| Error::message(format!("image processing task failed: {e}")))?
-}
-
-/// Maximum pages allowed in a single PDF Read call.
-const MAX_PDF_PAGES_PER_REQUEST: usize = 20;
-
-/// Parse a page-range string like `"1-5"`, `"3"`, or `"10-20"` into a
-/// `(start, end)` pair (1-indexed, inclusive).
-fn parse_page_range(raw: &str) -> Result<(usize, usize)> {
-    let raw = raw.trim();
-    if let Some((start_str, end_str)) = raw.split_once('-') {
-        let start: usize = start_str
-            .trim()
-            .parse()
-            .map_err(|_| Error::message(format!("invalid page range start: '{start_str}'")))?;
-        let end: usize = end_str
-            .trim()
-            .parse()
-            .map_err(|_| Error::message(format!("invalid page range end: '{end_str}'")))?;
-        if start == 0 || end == 0 {
-            return Err(Error::message(
-                "page numbers are 1-indexed (0 is not valid)",
-            ));
-        }
-        if start > end {
-            return Err(Error::message(format!(
-                "page range start ({start}) must be <= end ({end})"
-            )));
-        }
-        Ok((start, end))
-    } else {
-        let page: usize = raw
-            .parse()
-            .map_err(|_| Error::message(format!("invalid page number: '{raw}'")))?;
-        if page == 0 {
-            return Err(Error::message(
-                "page numbers are 1-indexed (0 is not valid)",
-            ));
-        }
-        Ok((page, page))
-    }
-}
-
-/// Read a PDF file and extract text by pages. Returns a structured
-/// payload with `kind: "pdf"`.
-///
-/// Uses the `pdf-extract` crate which is pure Rust and works on all
-/// platforms without native dependencies.
-async fn read_pdf(file_path: &str, pages: Option<&str>) -> Result<Value> {
-    require_absolute(file_path, "file_path")?;
-
-    let path = std::path::PathBuf::from(file_path);
-    if !fs::try_exists(&path).await.unwrap_or(false) {
-        return Ok(json!({
-            "kind": "not_found",
-            "file_path": file_path,
-            "error": "file does not exist",
-            "detail": "",
-        }));
-    }
-
-    let path_for_blocking = path.clone();
-    let pages_owned = pages.map(str::to_string);
-
-    tokio::task::spawn_blocking(move || -> Result<Value> {
-        let all_pages = match pdf_extract::extract_text_by_pages(&path_for_blocking) {
-            Ok(pages) => pages,
-            Err(e) => {
-                return Ok(json!({
-                    "kind": "pdf_error",
-                    "file_path": path_for_blocking.to_string_lossy(),
-                    "error": format!("failed to extract text from PDF: {e}"),
-                    "detail": "The file may be encrypted, corrupted, or use an unsupported PDF feature.",
-                }));
-            },
-        };
-
-        let total_pages = all_pages.len();
-
-        let (selected_pages, start_page, end_page) = if let Some(ref range_str) = pages_owned {
-            let (start, end) = parse_page_range(range_str)?;
-            let end = end.min(total_pages);
-            if start > total_pages {
-                return Ok(json!({
-                    "kind": "pdf",
-                    "file_path": path_for_blocking.to_string_lossy(),
-                    "total_pages": total_pages,
-                    "pages_returned": 0,
-                    "start_page": start,
-                    "end_page": end,
-                    "content": "",
-                    "error": format!("start page {start} exceeds total pages {total_pages}"),
-                }));
-            }
-            let page_count = end - start + 1;
-            if page_count > MAX_PDF_PAGES_PER_REQUEST {
-                return Err(Error::message(format!(
-                    "page range {start}-{end} spans {page_count} pages; maximum is {MAX_PDF_PAGES_PER_REQUEST} per request"
-                )));
-            }
-            let selected: Vec<&str> = all_pages[start - 1..end]
-                .iter()
-                .map(String::as_str)
-                .collect();
-            (selected, start, end)
-        } else {
-            // No pages specified: return all, capped at MAX_PDF_PAGES_PER_REQUEST.
-            let end = total_pages.min(MAX_PDF_PAGES_PER_REQUEST);
-            let selected: Vec<&str> = all_pages[..end]
-                .iter()
-                .map(String::as_str)
-                .collect();
-            (selected, 1, end)
-        };
-
-        // Build content with page markers.
-        let mut content = String::new();
-        for (idx, page_text) in selected_pages.iter().enumerate() {
-            let page_num = start_page + idx;
-            if !content.is_empty() {
-                content.push_str("\n\n");
-            }
-            content.push_str(&format!("--- Page {page_num} ---\n"));
-            content.push_str(page_text.trim());
-        }
-
-        Ok(json!({
-            "kind": "pdf",
-            "file_path": path_for_blocking.to_string_lossy(),
-            "total_pages": total_pages,
-            "pages_returned": selected_pages.len(),
-            "start_page": start_page,
-            "end_page": end_page,
-            "truncated": end_page < total_pages,
-            "content": content,
-        }))
-    })
-    .await
-    .map_err(|e| Error::message(format!("PDF extraction task failed: {e}")))?
-}
+use {image::is_image_extension, pdf::read_pdf};
 
 #[allow(clippy::unwrap_used, clippy::expect_used)]
 #[cfg(test)]
@@ -875,25 +683,25 @@ mod tests {
 
     #[test]
     fn parse_page_range_single_page() {
-        let (start, end) = parse_page_range("3").unwrap();
+        let (start, end) = pdf::parse_page_range("3").unwrap();
         assert_eq!((start, end), (3, 3));
     }
 
     #[test]
     fn parse_page_range_range() {
-        let (start, end) = parse_page_range("2-5").unwrap();
+        let (start, end) = pdf::parse_page_range("2-5").unwrap();
         assert_eq!((start, end), (2, 5));
     }
 
     #[test]
     fn parse_page_range_zero_rejected() {
-        assert!(parse_page_range("0").is_err());
-        assert!(parse_page_range("0-5").is_err());
+        assert!(pdf::parse_page_range("0").is_err());
+        assert!(pdf::parse_page_range("0-5").is_err());
     }
 
     #[test]
     fn parse_page_range_inverted_rejected() {
-        assert!(parse_page_range("5-2").is_err());
+        assert!(pdf::parse_page_range("5-2").is_err());
     }
 
     #[tokio::test]
